requirements.txt
langchain==1.2.10
langchain_core==1.2.15
langchain_google_genai==4.2.1
langgraph==1.0.9
python-dotenv==1.2.1
streamlit==1.54.0
langgraph-checkpoint-sqlite==3.0.3
langchain-community==0.4.1
ddgs==9.11.4
langchain-tavily==0.2.17
wikipedia==1.4.0